+++ src/main/java/com/ispp/heartforchange/service/GrantService.java
@@ -11,7 +11,7 @@ public interface GrantService {
 
 	GrantDTO getGrantById(Long id, String token) throws OperationNotAllowedException;
 	
-	Integer getTotalAmountAcceptedGrantsByOng(String token) throws OperationNotAllowedException;
+	Double getTotalAmountAcceptedGrantsByOng(String token) throws OperationNotAllowedException;
 
 	GrantDTO saveGrant(GrantDTO grantDTO, String token) throws OperationNotAllowedException;
